Development Do Some Changes In Backend and Frontend

+++ Event Managment Backend/loyalty-service/src/main/java/com/eventmanagement/loyalty/dto/LoyaltyResponseDTO.java
@@ -9,8 +9,8 @@ public class LoyaltyResponseDTO {
     private Long id;
     private Long attendeeId;
     private String attendeeName; // from Feign
-    private Integer pointsBalance;
-    private Integer totalPointsEarned;
+    private Double pointsBalance;
+    private Double totalPointsEarned;
     private String membershipTier;
     private String status;
     private LocalDateTime createdAt;
